60Hz Interrupt and Threading Issue Resolved
I decided to implement a timer interrupt at 60Hz to poll the size of
the terminal window. I ran into a really awesome threading issue once I
put this change into place.

The timer interrupt disables global interrupts while executing and then
attempts to write some bytes to the UART Tx buffer. UART transmission
is interrupt driven. If the Tx buffer is full, the UartTransmitByte
function will poll whether there is space in the buffer for the byte
that it wishes to send. Since interrupts are disabled the buffer will
never empty and the progam will spin on the full buffer (halt) :/

The solution is to unconditionally enable and disable interrupts in
the UartTransmitByte function to allow the buffer to empty out.

Threading issues: awesome :]