Map Tools, New Map Format
I decided to switch to a bit-packed binary format for the map rather
than run-length encoding. When maps started to grow the access times
to tiles at the end of the map became too high. A bit-packed format
allows for a constant access time at the cost of a slightly lower
packing ratio (in some cases).

I also created some crude tools to convert a map from a text file to
a C header file.