Improved Path Finding, Random Number Generator
I have improved the path finding algorithms and resolved a couple of
bugs. Each time step only moves one bug which allows the UI to be more
responsive. I first attempt to move the bot using the simple algorithm
and if that fails I will use the wave propagation algorithm. If that
fails, I move the bot to a random location on the map.

I implemented a random number generator based on the ADC. I have
debiased the results using the Von Neumann "Fair Coin" algorithm.