Tower Building, Inspection, Status Bar
I have implemented the ability to build towers. The bottom border is
designated as a status space to contain status of the most recent
action. It is not guaranteed to be painted there after the window
changes size.

You can now inspect bots by pressing the space bar and their remaining
health is displayed in the bottom bar.